Update prettier-plugin-organize-imports
The plugin is just a wrapper around TypeScript organizeImports function. The previous version of the plugin does not work with TypeScript 4.8, because of changed way of handling non-existing methods on a host object - the recent TypeScript implementations throws an error instead of silently continuing. The issue was resolved with the plugin providing an extended host object implementation with all necessary methods present.